pyproject.toml
[tool.poetry]
name = "youtube_monitor_action"
version = "0.1.0"
description = "Monitor for new video on YouTube Channel"
authors = ["mshafer1"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.4.1"
requests = "^2.26.0"
xmltodict = "^0.12.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
pytest-snapshot = "^0.6.1"
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"

[tool.poetry.scripts]
wait-for-videos = "youtube_monitor_action.__main__:main"
